+++ OffOff_iOS/OffOff_iOS/Commons/ContentSizedTableView.swift
@@ -1,0 +1,21 @@
+//
+//  ContentSizedTableView.swift
+//  OffOff_iOS
+//
+//  Created by Lee Nam Jun on 2021/10/13.
+//
+
+import UIKit
+
+final class ContentSizedTableView: UITableView {
+    override var contentSize:CGSize {
+        didSet {
+            invalidateIntrinsicContentSize()
+        }
+    }
+
+    override var intrinsicContentSize: CGSize {
+        layoutIfNeeded()
+        return CGSize(width: UIView.noIntrinsicMetric, height: contentSize.height)
+    }
+}
