+++ OffOff_iOS/OffOff_iOS/Views/PostView.swift
@@ -146,16 +146,3 @@ class PostView: UIScrollView {
         contentTextView.text = post.content
     }
 }
-
-final class ContentSizedTableView: UITableView {
-    override var contentSize:CGSize {
-        didSet {
-            invalidateIntrinsicContentSize()
-        }
-    }
-
-    override var intrinsicContentSize: CGSize {
-        layoutIfNeeded()
-        return CGSize(width: UIView.noIntrinsicMetric, height: contentSize.height)
-    }
-}
